+++ src/monorepo_pnpm__dependencies__update__cli/index.d.ts
@@ -1,0 +1,1 @@
+export declare function monorepo_pnpm__dependencies__update__cli():Promise<void>